+++ src/components/BoutonsSociaux.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 function BoutonsSociaux(props) {
-  const url = "http://10.0.51.241:8800";
+  const url = "http://192.168.1.80:8800";
 
   return (
     <div>
