+++ src/pages/dashboard/DashBoard.jsx
@@ -5,9 +5,8 @@ import NavigBar from '../../components/NavigBar';
 import Footer from '../../components/Footer';
 import { COLORS } from '../../utils/colors';
 import API from '../../utils/API';
-import SessionPostChild from '../session/SessionPostChild';
-import { stringToDate } from '../../utils/dates';
 import { fetchSeancesData } from '../../utils/seance';
+import DisplaySeancesPost from '../../components/DisplaySeancesPost';
 
 const DashBoard = () => {
     const { width } = useWindowDimensions();
@@ -63,8 +62,6 @@ const DashBoard = () => {
         </div>
     }
 
-    const backgroundColors = ["#9C005D", "#9C1B00", "#9B0000", "#8B009C", "#9C3600"];
-
     return (
         <div style={{ backgroundColor: COLORS.PAGE_BACKGROUND }}>
             <div className="page-container">
@@ -118,32 +115,7 @@ const DashBoard = () => {
                     </div>
 
                     {/* SEANCES */}
-                    <div className='basic-flex popInElement' style={{ flexDirection: 'column', gap: '40px', alignItems: 'center' }}>
-                        {seances && seances.length > 0 ? (
-                            seances.map((seance, index) => (
-                                <div className="session-post" style={
-                                    width < 400 ? { padding: '5px', margin: "20px 0 0 0" } :
-                                        width < 550 ? { padding: '10px', margin: "20px 10px 0 10px" } :
-                                            { padding: '20px', margin: "20px 20px 0 20px" }}>
-                                    <SessionPostChild
-                                        key={seance._id} // Always provide a key when rendering lists
-                                        user={seance.user}
-                                        postTitle={seance.title ? seance.title : "N/A"}
-                                        postDescription={seance.description}
-                                        selectedName={seance.name}
-                                        selectedDate={stringToDate(seance.date)}
-                                        selectedExercices={seance.exercices}
-                                        stats={seance.stats ? seance.stats : { nSets: "N/A", nReps: "N/A", intervalReps: "N/A", totalWeight: "N/A", intervalWeight: "N/A" }}
-                                        backgroundColors={backgroundColors}
-                                        recordSummary={seance.recordSummary ? seance.recordSummary : []}
-                                        editable={false}
-                                    />
-                                </div>
-                            ))
-                        ) : (
-                            <div>Aucune séances</div>
-                        )}
-                    </div>
+                    <DisplaySeancesPost seances={seances} />
                 </div>
